+++ include/dialogs.h
@@ -184,6 +184,10 @@ extern struct UIItem diaPadEmuInfo[];
 #endif
 
 extern struct UIItem diaNetConfig[];
+
+extern struct UIItem diaModesConfig[];
+extern struct UIItem diaVMCConfig[];
+
 extern struct UIItem diaCompatConfig[];
 extern struct UIItem diaUIConfig[];
 extern struct UIItem diaAudioConfig[];
