+++ include/menusys.h
@@ -98,9 +98,11 @@ void menuSaveConfig();
 void menuRenderMain();
 void menuRenderMenu();
 void menuRenderInfo();
+void menuRenderGameMenu();
 void menuHandleInputMain();
 void menuHandleInputMenu();
 void menuHandleInputInfo();
+void menuHandleInputGameMenu();
 
 // Sets the selected item if it is found in the menu list
 void menuSetSelectedItem(menu_item_t *item);
